Make germqtt a regular user
To use germqtt it needs to ssh into gerrit to get the event stream.
Initiating an ssh connection by default assumes there is a home
directory for the user, and that it can use that to store stuff. To
make things simpler this commit changes the germqtt user to be a
regular user instead of a system user to facilitate this.